// web/src/shared/api/generated/model/multiUnitExamOverviewResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { InProgressExamResult } from './inProgressExamResult';
import type { MultiUnitExamUnitOptionResult } from './multiUnitExamUnitOptionResult';

export interface MultiUnitExamOverviewResult {
  subjectId: string;
  subjectName: string;
  units: MultiUnitExamUnitOptionResult[];
  /** @items.pattern ^-?(?:0|[1-9]\d*)$ */
  sizes: (number | string)[];
  inProgressExam: null | InProgressExamResult;
}