// web/src/shared/api/generated/model/multiUnitExamPreviewResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ExamBlueprintSummaryResult } from './examBlueprintSummaryResult';
import type { MultiUnitExamUnitShareResult } from './multiUnitExamUnitShareResult';

export interface MultiUnitExamPreviewResult {
  subjectId: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  size: number | string;
  blueprint: ExamBlueprintSummaryResult;
  isAvailable: boolean;
  units: MultiUnitExamUnitShareResult[];
}